# CASEProvider/public/openapi-spec.json
{
    "swagger" : "2.0",
    "info" : {
        "version" : "1.1",
        "title" : "Competencies and Academic Standards Exchange (CASE) Service OpenAPI (JSON) Definition",
        "description" : "The Competencies and Academic Standards Exchange (CASE) Service enables the exchange of data between a Competency Records Service Provider and the consumers of the associated data. This service has been described using the 1EdTech Model Driven Specification development approach - this being the Platform Specific Model (PSM) of the service.",
        "termsOfService" : "Permission is granted to all parties to use excerpts from this document as needed in producing requests for proposals. The limited permissions granted above are perpetual and will not be revoked by 1EdTech or its successors or assigns. THIS API DEFINITION IS BEING OFFERED WITHOUT ANY WARRANTY WHATSOEVER, AND IN PARTICULAR, ANY WARRANTY OF NONINFRINGEMENT IS EXPRESSLY DISCLAIMED. ANY USE OF THIS SPECIFICATION SHALL BE MADE ENTIRELY AT THE IMPLEMENTERS OWN RISK, AND NEITHER THE  CONSORTIUM, NOR ANY OF ITS MEMBERS OR SUBMITTERS, SHALL HAVE ANY LIABILITY WHATSOEVER TO ANY IMPLEMENTER OR THIRD PARTY FOR ANY DAMAGES OF ANY NATURE WHATSOEVER,  DIRECTLY OR INDIRECTLY, ARISING FROM THE USE OF THIS API DEFINITION.",
        "contact" : {
            "name" : "Vice President of Operations, 1EdTech",
            "url" : "https://www.1edtech.org",
            "email" : "operations@1edtech.org"
        },
        "license" : {
...
    "host" : "www.imsglobal.org",
    "basePath" : "/ims/case/v1p1",
    "schemes" : [ "http","https" ],
...
    "paths" : {
...
        "/CFDocuments" : {
            "get" : {
                "operationId" : "getAllCFDocuments",
                "summary" : "The REST read request message for the getAllCFDocuments() API call.",
                "tags" : [ "DocumentsManager" ],
                "description" : "This is a request to the Service Provider to provide all of the Competency Framework Documents. ",
                "parameters" : [
                    {
                        "name" : "limit",
                        "in" : "query",
                        "description" : "This is used as part of the data pagination mechanism to control the download rate of data. The 'limit' defines the download segmentation value i.e. the maximum number of records to be contained in the response. The form of implementation is described in the corresponding binding document(s).",
                        "required" : false,
                        "type" : "integer",
                        "format" : "int32",
                        "minimum" : 1,
                        "allowEmptyValue" : false
                    },
                    {
                        "name" : "offset",
                        "in" : "query",
                        "description" : "This is used as part of the data pagination mechanism to control the download rate of data. The 'offset' is the number of the first record to be supplied in the segmented response message. The form of implementation is described in the corresponding binding document(s).",
                        "required" : false,
                        "type" : "integer",
                        "format" : "int32",
                        "default" : 0,
                        "minimum" : 0,
                        "allowEmptyValue" : false
                    },
                    {
                        "name" : "sort",
                        "in" : "query",
                        "description" : "This is used as part of the sorting mechanism to be use by the service provider. The 'sort' identifies the sort criteria to be used for the records in the response message. Use with the orderBy parameter. The form of implementation is described in the corresponding binding document(s).",
                        "required" : false,
                        "type" : "string",
                        "allowEmptyValue" : false
                    },
                    {
                        "name" : "orderBy",
                        "in" : "query",
                        "description" : "This is used as part of the sorting mechanism to be use by the service provider. This defines the form of ordering for response to the sorted request i.e. ascending (asc) or descending (desc). The form of implementation is described in the corresponding binding document(s).",
                        "required" : false,
                        "type" : "string",
                        "enum" : [ "asc","desc" ],
                        "allowEmptyValue" : false
                    },
                    {
                        "name" : "filter",
                        "in" : "query",
                        "description" : "This is used for the data filtering mechanism to be applied by the service provider. It defines the filtering rules to be applied when identifying the records to be supplied in the response message. The form of implementation is described in the corresponding binding document(s).",
                        "required" : false,
                        "type" : "string",
                        "allowEmptyValue" : false
                    },
                    {
                        "name" : "fields",
                        "in" : "query",
                        "description" : "This is used as part of the field selection mechanism to be applied by the service provider. This identifies the range of fields that should be supplied in the response message. The form of implementation is described in the corresponding binding document(s).",
                        "required" : false,
                        "type" : "array",
                        "items" : {
                            "type" : "string"
                        },
                        "collectionFormat" : "csv",
                        "allowEmptyValue" : false
                    }
                ],
                "x-1edtech-confidentiality" : "unrestricted",
                "responses" : {
                    "200" : {
                        "description" : "This is the response when the request has been completed successfully. It is the set of CFDocuments from the service provider.",
                        "schema" : {
                            "$ref" : "#/definitions/CFDocumentSetDType"
                        },
                        "headers" : {
                            "X-Total-Count" : {
                                "description" : "The total number of resources that are available to be returned",
                                "type" : "integer"
                            }
                        }
                    },
                    "400" : {
                        "description" : "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "401" : {
                        "description" : "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "403" : {
                        "description" : "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "404" : {
                        "description" : "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "429" : {
                        "description" : "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "500" : {
                        "description" : "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "default" : {
                        "description" : "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    }
                }
            }
        },
        "/CFDocuments/{sourcedId}" : {
            "get" : {
                "operationId" : "getCFDocument",
                "summary" : "The REST read request message for the getCFDocument() API call.",
                "tags" : [ "DocumentsManager" ],
                "description" : "This is a request to the service provider to provide the information for the specific Competency Framework Document. If the identified record cannot be found then the 'unknownobject' status code must be reported.",
                "parameters" : [
                    {
                        "name" : "sourcedId",
                        "in" : "path",
                        "description" : "The UUID that identifies the Competency Framework Document that is to be read from the service provider.",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "x-1edtech-confidentiality" : "unrestricted",
                "responses" : {
                    "200" : {
                        "description" : "This is the response when the request has been completed successfully. It is the CFDocument from the service provider.",
                        "schema" : {
                            "$ref" : "#/definitions/CFDocumentDType"
                        }
                    },
                    "400" : {
                        "description" : "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "401" : {
                        "description" : "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "403" : {
                        "description" : "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "404" : {
                        "description" : "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "429" : {
                        "description" : "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "500" : {
                        "description" : "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "default" : {
                        "description" : "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    }
                }
            }
        },
...
    },
    "definitions" : {
...
        "CFDocumentDType" : {
            "description" : "The container for the data about a competency framework document (CFDocument) when exchanged outside of the context of a CFPackage. A CFDocument is the root for the creation of a learning standard/competency.",
            "type" : "object",
            "properties" : {
                "identifier" : {
                    "description" : "Model Primitive Datatype = NormalizedString. An unambiguous, synthetic, globally unique identifier for the CFDocument. This is the primary way in which the exchange identification is achieved.",
                    "type" : "string",
                    "pattern" : "[0-9a-f]{8}-[0-9a-f]{4}-[1-5]{1}[0-9a-f]{3}-[8-9a-b]{1}[0-9a-f]{3}-[0-9a-f]{12}"
                },
                "uri" : {
                    "description" : "Model Primitive Datatype = AnyURI. An unambiguous reference to the CFDocument using a network-resolvable URI.",
                    "type" : "string",
                    "format" : "uri"
                },
                "frameworkType" : {
                    "description" : "Model Primitive Datatype = NormalizedString. This is a new attribute added in version 1.1. This attribute allows framework creators to indicate what type of framework this is, for example course codes. In CASE 1.1. the only predefined type of framework is 'CourseCodes'. Other standardized framework types will be defined.",
                    "type" : "string"
                },
                "caseVersion" : {
                    "description" : "This is a new attribute added in version 1.1. Denotes the version of the CFDocument. If present it MUST have a value of '1.1'.",
                    "type" : "string",
                    "enum" : [ "1.1" ]
                },
                "creator" : {
                    "description" : "Model Primitive Datatype = NormalizedString. The entity with authority that promulgates the competency framework. This is the entity that authorized or created the competency framework. It could be an education agency, higher education institution, professional body. It is the owner of the competency framework.",
                    "type" : "string"
                },
                "title" : {
                    "description" : "Model Primitive Datatype = NormalizedString. The title of the CFDocument.",
                    "type" : "string"
                },
                "lastChangeDateTime" : {
                    "description" : "Model Primitive Datatype = DateTime. A system generated timestamp of the most recent change to this record. This conforms to ISO 8601 dateTime definition [ISO 8601].",
                    "type" : "string",
                    "format" : "date-time"
                },
                "officialSourceURL" : {
                    "description" : "Model Primitive Datatype = AnyURI. The URL link to the formal citation of the original achievement standards document created for humans.",
                    "type" : "string",
                    "format" : "uri"
                },
                "publisher" : {
                    "description" : "Model Primitive Datatype = NormalizedString. The entity responsible for making the learning standards document available.",
                    "type" : "string"
                },
                "description" : {
                    "description" : "Model Primitive Datatype = String. A human readable description of the CFDocument. In Version 1.1 the data-type has been changed from NormalizedString.",
                    "type" : "string"
                },
                "subject" : {
                    "description" : "Model Primitive Datatype = NormalizedString. The topic or academic subject of the Document (this should use some appropriate K-12, HE, etc. based vocabulary).",
                    "type" : "array",
                    "minItems" : 0,
                    "items" : {
                        "type" : "string"
                    }
                },
                "subjectURI" : {
                    "description" : "A set of Link URIs denoting the set of subjects for the document as designated by the promulgating agency. ",
                    "type" : "array",
                    "minItems" : 0,
                    "items" : {
                        "$ref" : "#/definitions/LinkURIDType"
                    }
                },
                "language" : {
                    "description" : "Model Primitive Datatype = Language. The default language of the text used for the content in the learning standard document (use ISO 639-2 [ISO 639]).",
                    "type" : "string"
                },
                "version" : {
                    "description" : "Model Primitive Datatype = NormalizedString. Defines the revision of the document (the nature of this versioning is an implementation issue).",
                    "type" : "string"
                },
                "adoptionStatus" : {
                    "description" : "Model Primitive Datatype = NormalizedString. The publication status of the document.",
                    "type" : "string"
                },
                "statusStartDate" : {
                    "description" : "Model Primitive Datatype = Date. The date the CFDocument status started.",
                    "type" : "string",
                    "format" : "date"
                },
                "statusEndDate" : {
                    "description" : "Model Primitive Datatype = Date. The date the CFDocument status ended or changed to another status.",
                    "type" : "string",
                    "format" : "date"
                },
                "licenseURI" : {
                    "$ref" : "#/definitions/LinkURIDType"
                },
                "notes" : {
                    "description" : "Model Primitive Datatype = String. Any text used to comment on the published CFDocument.",
                    "type" : "string"
                },
                "extensions" : {
                    "$ref" : "#/definitions/CFDocumentExtensionDType"
                },
                "CFPackageURI" : {
                    "$ref" : "#/definitions/LinkURIDType"
                }
            },
            "required" : [ "identifier","uri","creator","title","lastChangeDateTime","CFPackageURI" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
        "CFDocumentExtensionDType" : {
            "description" : "The container for the proprietary extensions to the CFDocument class.",
            "type" : "object",
            "properties" : {

            },
            "x-1edtech-privacy" : [ "extensions" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : true
        },
        "CFDocumentSetDType" : {
            "description" : "This is the container for a collection of CFDocuments. There must be at least one CFDocument.",
            "type" : "object",
            "properties" : {
                "CFDocuments" : {
                    "description" : "This is the data for a single CFDocument within the CFDocumentSet container.",
                    "type" : "array",
                    "minItems" : 1,
                    "items" : {
                        "$ref" : "#/definitions/CFDocumentDType"
                    }
                }
            },
            "required" : [ "CFDocuments" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
...
        "LinkURIDType" : {
            "description" : "A container for the information that is used to achieve the link data reference.",
            "type" : "object",
            "properties" : {
                "title" : {
                    "description" : "Model Primitive Datatype = NormalizedString. A human readable title for the associated object.",
                    "type" : "string"
                },
                "identifier" : {
                    "description" : "Model Primitive Datatype = NormalizedString. An unambiguous, synthetic, globally unique identifier (UUID) for the associated object.",
                    "type" : "string",
                    "pattern" : "[0-9a-f]{8}-[0-9a-f]{4}-[1-5]{1}[0-9a-f]{3}-[8-9a-b]{1}[0-9a-f]{3}-[0-9a-f]{12}"
                },
                "uri" : {
                    "description" : "Model Primitive Datatype = AnyURI. A network-resolvable URI pointing to the authoritative reference for the associated object.",
                    "type" : "string",
                    "format" : "uri"
                }
            },
            "required" : [ "title","identifier","uri" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
        "imsx_CodeMinorDType" : {
            "description" : "This is the container for the set of code minor status codes reported in the responses from the Service Provider.",
            "type" : "object",
            "properties" : {
                "imsx_codeMinorField" : {
                    "description" : "Each reported code minor status code.",
                    "type" : "array",
                    "minItems" : 1,
                    "items" : {
                        "$ref" : "#/definitions/imsx_CodeMinorFieldDType"
                    }
                }
            },
            "required" : [ "imsx_codeMinorField" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
        "imsx_CodeMinorFieldDType" : {
            "description" : "This is the container for a single code minor status code.",
            "type" : "object",
            "properties" : {
                "imsx_codeMinorFieldName" : {
                    "description" : "Model Primitive Datatype = NormalizedString. This should contain the identity of the system that has produced the code minor status code report.",
                    "type" : "string"
                },
                "imsx_codeMinorFieldValue" : {
                    "description" : "The code minor status code (this is a value from the corresponding enumerated vocabulary).",
                    "type" : "string",
                    "enum" : [ "forbidden","fullsuccess","internal_server_error","invalid_selection_field","invalid_sort_field","invalid_uuid","server_busy","unauthorised_request","unknownobject" ]
                }
            },
            "required" : [ "imsx_codeMinorFieldName","imsx_codeMinorFieldValue" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
        "imsx_StatusInfoDType" : {
            "description" : "This is the container for the status code and associated information returned within the HTTP messages received from the Service Provider. For the CASE service this object will only be returned to provide information about a failed request i.e. it will NOT be in the payload for a successful request. See Appendix B for further information on the interpretation of the information contained within this class",
            "type" : "object",
            "properties" : {
                "imsx_codeMajor" : {
                    "description" : "The code major value (from the corresponding enumerated vocabulary). See Appendix B for further information on the interpretation of this set of codes.",
                    "type" : "string",
                    "enum" : [ "failure","processing","success","unsupported" ]
                },
                "imsx_severity" : {
                    "description" : "The severity value (from the corresponding enumerated vocabulary). See Appendix B for further information on the interpretation of this set of codes.",
                    "type" : "string",
                    "enum" : [ "error","status","warning" ]
                },
                "imsx_description" : {
                    "description" : "Model Primitive Datatype = String. A human readable description supplied by the entity creating the status code information.",
                    "type" : "string"
                },
                "imsx_codeMinor" : {
                    "$ref" : "#/definitions/imsx_CodeMinorDType"
                }
            },
            "required" : [ "imsx_codeMajor","imsx_severity" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        }
    },
    "consumes" : [ "application/json" ],
    "produces" : [ "application/json" ]
}